try split
